Add image proxy
