+++ homeassistant/components/sonos/exception.py
@@ -1,0 +1,6 @@
+"""Sonos specific exceptions."""
+from homeassistant.components.media_player.errors import BrowseError
+
+
+class UnknownMediaType(BrowseError):
+    """Unknown media type."""
